+++ language/en/common.php
@@ -79,10 +79,9 @@ $lang = array_merge($lang, [
 		2	=> 'Currently <strong>%1$d</strong> items are listed in the database.',
 	],
 	'MOT_TZV_COUNTRY_EINTRAG'			=> 'Registered countries',
-	'MOT_TZV_MAINNEWS_INFO'				=> '&bull;&nbsp;&nbsp;Added an OSM search (Nominatim) to the map.<br>
-											&bull;&nbsp;&nbsp;Added a data transfer to komoot.de to enable route planning for hikers and bikers (opens in a new tab or window).<br>
-											&bull;&nbsp;&nbsp;komoot.de will be called with a localisation variable to set the users language if this language is supported (e.g. if user uses English komoot.de will open in English).<br>
-											&bull;&nbsp;&nbsp;Added the ability to create a new item by right-clicking onto the desired map location and then be redirected to the input form with this location`s coordinates.<br>',
+	'MOT_TZV_MAINNEWS_INFO'				=> '&bull;&nbsp;&nbsp;Displaying the items in different map overlays named after the used categories.<br>
+											&bull;&nbsp;&nbsp;Overlays can be switched on and off individually with the layer control element in the map`s upright corner.<br>
+											&bull;&nbsp;&nbsp;Displaying the map`s distance scale in metric and imperial units.<br>',
 	'MOT_TZV_NUTZUNG_MAPS'				=> 'Google Maps Terms of Use',
 
 	// Map